Move exceptions to new module
