+++ gator/orchestrate.py
@@ -11,10 +11,11 @@ from gator import description
 from gator import leave
 from gator import report
 
-# pylint: disable=unused-import
-from gator import display  # noqa: F401
-from gator import invoke  # noqa: F401
-from gator import run  # noqa: F401
+from gator.exceptions import (
+    InvalidSystemArgumentsError,
+    InvalidCheckArgumentsError,
+    InvalidCheckError,
+)
 
 # define the name of this module
 ORCHESTRATE = sys.modules[__name__]
@@ -27,23 +28,6 @@ REPORT = sys.modules[constants.modules.Report]
 
 # define the format for the output of the checks
 OUTPUT_TYPE = getattr(REPORT, constants.outputs.Text)
-
-
-class InvalidSystemArgumentsError(ValueError):
-    """The system arguments are invalid."""
-
-    def __init__(self, arguments):
-        """Initialize the InvalidSystemArgumentsError."""
-        self.arguments = arguments
-
-
-class InvalidCheckError(InvalidSystemArgumentsError):
-    """The check is invalid."""
-
-    def __init__(self, arguments, check_name) -> None:
-        """Initialize the InvalidCheckError."""
-        super().__init__(arguments)
-        self.check_name = check_name
 
 
 def get_welcome_actions():
